Parsing into data struct mostly done

+++ mjgwin/dims.h
@@ -10,6 +10,6 @@ typedef enum dim {
   num_dims
 } dim_t;
 
-typedef int8_t pair_t[num_dims];
+typedef int16_t pair_t[num_dims];
 
 #endif
